cleaning up compiler warnings --- mostly signed/unsigned int comparisons.  something is broken in testsuite when running with 2 procs. single proc is ok.

+++ src/madness/chem/ccpairfunction.h
@@ -699,7 +699,7 @@ public:
                                            const std::array<int, LDIM>& v1) {
         auto a012=std::array<int,LDIM>();
         auto a345=std::array<int,LDIM>();
-        for (int i=0; i<LDIM; ++i) {
+        for (size_t i=0; i<LDIM; ++i) {
             a012[i]=i;
             a345[i]=i+LDIM;
         }
